+++ app/controllers/panels_controller.rb
@@ -17,14 +17,14 @@ class PanelsController < IndexController
     summary = {
       name:                           panel.name,
       nation:                         panel.nation,
-      number_of_participated_segment:
-        PerformedSegment.joins(:officials).where("officials.panel": panel).count,
-      number_of_scores_judged:
-        Score.joins(performed_segment: [:officials]).where("officials.panel_id": panel.id).count,
+#      number_of_participated_segment:
+#        PerformedSegment.joins(:officials).where("officials.panel": panel).count,
+#      number_of_scores_judged:
+#        Score.joins(performed_segment: [:officials]).where("officials.panel_id": panel.id).count,
     }
 
     columns = [:competition_name, :category_name, :segment_name, :number]
-    records = Official.where(panel: panel).includes(performed_segment: [:competition, :category, :segment])
+    records = panel.officials.includes(performed_segment: [:competition, :category, :segment])
     participated_segments_datatable =
       AjaxDatatables::Datatable.new(self).records(records).columns(columns)