+++ app/models/competition.rb
@@ -4,9 +4,9 @@ class Competition < ApplicationRecord
   alias_attribute :competition_name, :name
 
   ## relations
+  has_many :performed_segments, dependent: :destroy
   has_many :category_results, dependent: :destroy
   has_many :scores, dependent: :destroy
-  has_many :performed_segments, dependent: :destroy
 
   ## validations
   validates :country, allow_nil: true, format: { with: /\A[A-Z][A-Z][A-Z]\Z/ }