+++ app/models/performed_segment.rb
@@ -1,7 +1,6 @@
 class PerformedSegment < ApplicationRecord
   has_many :scores, dependent: :nullify
   has_many :officials, dependent: :destroy
-  validates  :starting_time, presence: true
 
   belongs_to :competition
   belongs_to :category
@@ -10,4 +9,6 @@ class PerformedSegment < ApplicationRecord
   delegate :competition_name, to: :competition
   delegate :category_name, to: :category
   delegate :segment_name, to: :segment
+
+  validates  :starting_time, presence: true
 end