+++ app/updaters/competition_updater.rb
@@ -151,12 +151,12 @@ class CompetitionUpdater
             score.attributes = slice_common_attributes(score, parsed)
                                .merge(
                                  skater: skater,
-                                 performed_segment: ps,
+                                 #performed_segment: ps,
                                  date: ps.starting_time.to_date
                                )
 
             if relevant_cr
-              score.update(category_result: relevant_cr)
+              #score.update(category_result: relevant_cr)
               score.save!           ## need to save here
               relevant_cr.update(segment.segment_type => score)
             end